+++ inventory/__init__.py
@@ -1,1 +1,6 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 from . import filter_tools, get_stats, get_tools, util