+++ inventory/util.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 """Various util functions for inventory collection, filtering, and stats getting."""
 
 import logging