+++ website/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT