+++ website/pages/1_👤_Deep_Dive_-_User_Interaction_Analysis.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 """Create Streamlit web app sub-page to visualise tool user data.
 
 (C) Open Energy Transition (OET)